requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
sqlalchemy>=2.0.0
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
groq
openai
hindsight-client
requests
python-multipart
